made changes for advanced crafting table and addon additions
Made it so if you were to place an item in a advanced crafting table slot where it's part of the gui, then it will return the item in the form of a item entity on top of the block.  # Bugged and needs to be fixed.

Added Furniture addon (Work-in-progress)